# Remove brain farts