--- a/WalletWasabi/Blockchain/Keys/HdPubKeyCache.cs
+++ b/WalletWasabi/Blockchain/Keys/HdPubKeyCache.cs
@@ -1,31 +1,28 @@
 using System.Collections;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using NBitcoin;
 
 namespace WalletWasabi.Blockchain.Keys;
 
-public class HdPubKeyCache : IEnumerable<HdPubKey>
+public class HdPubKeyCache : IEnumerable<HdPubKeyInfo>
 {
-	private Dictionary<Script, HdPubKey> HdPubKeysByScript { get; } = new();
-	private HashSet<HdPubKey> HdPubKeys { get; } = new();
-	private Dictionary<KeyPath, ScriptBytesHdPubKeyPair> ScriptBytesHdPubKeyPairByKeyPath { get; } = new();
+	private Dictionary<Script, HdPubKey> _hdPubKeyIndexedByScriptPubKey = new(1_000);
+	private HdPubKeyInfo[] _hdPubKeyInfos = new HdPubKeyInfo[1_000];
+	private int _hdKeyCount = 0;
 
-	private HdPubKeyGlobalView Snapshot =>
-		new(this.ToImmutableList());
+	private ArraySegment<HdPubKeyInfo> Snapshot =>
+		new (_hdPubKeyInfos, 0, _hdKeyCount);
 
-	public IEnumerable<SynchronizationInfos> GetSynchronizationInfos()
-	{
-		return ScriptBytesHdPubKeyPairByKeyPath.Select(x => new SynchronizationInfos(x.Key, x.Value));
-	}
-	
+	public IEnumerable<HdPubKey> HdPubKeys =>
+		Snapshot.Select(x => x.HdPubKey);
+
 	public bool TryGetPubKey(Script destination, [NotNullWhen(true)] out HdPubKey? hdPubKey) =>
-		HdPubKeysByScript.TryGetValue(destination, out hdPubKey);
+		_hdPubKeyIndexedByScriptPubKey.TryGetValue(destination, out hdPubKey);
 
 	public HdPubKeyPathView GetView(KeyPath keyPath) =>
-		Snapshot.GetChildKeyOf(keyPath);
+		new(HdPubKeys.Where(x => x.FullKeyPath.Parent == keyPath));
 
 	public IEnumerable<HdPubKey> AddRangeKeys(IEnumerable<HdPubKey> keys)
 	{
@@ -39,20 +36,22 @@ public class HdPubKeyCache : IEnumerable<HdPubKey>
 
 	public void AddKey(HdPubKey hdPubKey, ScriptPubKeyType scriptPubKeyType)
 	{
-		var scriptPubKey = hdPubKey.PubKey.GetScriptPubKey(scriptPubKeyType);
-		HdPubKeysByScript.AddOrReplace(scriptPubKey, hdPubKey);
-		ScriptBytesHdPubKeyPairByKeyPath.AddOrReplace(hdPubKey.FullKeyPath, new ScriptBytesHdPubKeyPair(scriptPubKey.ToCompressedBytes(), hdPubKey));
-		HdPubKeys.Add(hdPubKey);
+		if (_hdPubKeyInfos.Length == _hdKeyCount)
+		{
+			Array.Resize(ref _hdPubKeyInfos, 2 * _hdKeyCount);
+		}
+
+		var info = new HdPubKeyInfo(hdPubKey, scriptPubKeyType);
+		_hdPubKeyInfos[_hdKeyCount] = info;
+		_hdPubKeyIndexedByScriptPubKey[info.ScriptPubKey] = info.HdPubKey;
+		_hdKeyCount++;
 	}
 
-	public IEnumerator<HdPubKey> GetEnumerator() =>
-		HdPubKeys
-		.OrderBy(x => x.Index)
-		.GetEnumerator();
+	public IEnumerator<HdPubKeyInfo> GetEnumerator() =>
+		Snapshot
+			.OrderBy(x => x.HdPubKey.Index)
+			.GetEnumerator();
 
 	IEnumerator IEnumerable.GetEnumerator() =>
 		GetEnumerator();
-
-	public record ScriptBytesHdPubKeyPair(byte[] ScriptBytes, HdPubKey HdPubKey);
-	public record SynchronizationInfos(KeyPath KeyPath, ScriptBytesHdPubKeyPair ScriptBytesHdPubKeyPair);
 }