--- a/WalletWasabi/Blockchain/Keys/HdPubKeyPathView.cs
+++ b/WalletWasabi/Blockchain/Keys/HdPubKeyPathView.cs
@@ -7,12 +7,12 @@ namespace WalletWasabi.Blockchain.Keys;
 
 public class HdPubKeyPathView : IEnumerable<HdPubKey>
 {
-	internal HdPubKeyPathView(ImmutableList<HdPubKey> hdPubKeys)
+	internal HdPubKeyPathView(IEnumerable<HdPubKey> hdPubKeys)
 	{
 		Keys = hdPubKeys;
 	}
 
-	protected ImmutableList<HdPubKey> Keys { get; }
+	protected IEnumerable<HdPubKey> Keys { get; }
 	public IEnumerable<HdPubKey> CleanKeys => GetKeysByState(KeyState.Clean);
 	public IEnumerable<HdPubKey> LockedKeys => GetKeysByState(KeyState.Locked);
 	public IEnumerable<HdPubKey> UsedKeys => GetKeysByState(KeyState.Used);