--- a/WalletWasabi/Blockchain/Keys/KeyManager.cs
+++ b/WalletWasabi/Blockchain/Keys/KeyManager.cs
@@ -113,7 +113,7 @@ public class KeyManager
 	private void OnSerializingMethod(StreamingContext context)
 	{
 		HdPubKeys.Clear();
-		HdPubKeys.AddRange(HdPubKeyCache);
+		HdPubKeys.AddRange(HdPubKeyCache.HdPubKeys);
 		MinGapLimit = Math.Max(SegwitExternalKeyGenerator.MinGapLimit, TaprootExternalKeyGenerator?.MinGapLimit ?? 0);
 	}
 
@@ -412,8 +412,8 @@ public class KeyManager
 			AssertCleanKeysIndexed();
 			return wherePredicate switch
 			{
-				null => HdPubKeyCache.ToList(),
-				_ => HdPubKeyCache.Where(wherePredicate).ToList()
+				null => HdPubKeyCache.Select(x => x.HdPubKey),
+				_ => HdPubKeyCache.Select(x => x.HdPubKey).Where(wherePredicate)
 			};
 		}
 	}
@@ -432,11 +432,12 @@ public class KeyManager
 	/// It's unsafe because it doesn't assert that the GapLimit is respected.
 	/// GapLimit should be enforced whenever a transaction is discovered.
 	/// </summary>
-	public IEnumerable<HdPubKeyCache.SynchronizationInfos> UnsafeGetSynchronizationInfos()
+	public record ScriptPubKeySpendingInfo(byte[] CompressedScriptPubKey, Height? LatestSpendingHeight);
+	public IEnumerable<ScriptPubKeySpendingInfo> UnsafeGetSynchronizationInfos()
 	{
 		lock (CriticalStateLock)
 		{
-			return HdPubKeyCache.GetSynchronizationInfos();
+			return HdPubKeyCache.Select(x => new ScriptPubKeySpendingInfo(x.CompressedScriptPubKey, x.HdPubKey.LatestSpendingHeight));
 		}
 	}
 
@@ -450,31 +451,22 @@ public class KeyManager
 
 	public IEnumerable<ExtKey> GetSecrets(string password, params Script[] scripts)
 	{
-		return GetSecretsAndPubKeyPairs(password, scripts).Select(x => x.secret);
+		return GetSecretsAndPubKeyPairs(password, scripts);
 	}
 
-	public IEnumerable<(ExtKey secret, HdPubKey pubKey)> GetSecretsAndPubKeyPairs(string password, params Script[] scripts)
+	public IEnumerable<ExtKey> GetSecretsAndPubKeyPairs(string password, params Script[] scripts)
 	{
 		ExtKey extKey = GetMasterExtKey(password);
-		var extKeysAndPubs = new List<(ExtKey secret, HdPubKey pubKey)>();
+		var extKeysAndPubs = new List<ExtKey>();
 
 		lock (CriticalStateLock)
 		{
 			foreach (HdPubKey key in GetKeys(x =>
 				scripts.Contains(x.P2wpkhScript)
-				|| scripts.Contains(x.P2shOverP2wpkhScript)
-				|| scripts.Contains(x.P2pkhScript)
-				|| scripts.Contains(x.P2pkScript)
 				|| scripts.Contains(x.P2Taproot)))
 			{
 				ExtKey ek = extKey.Derive(key.FullKeyPath);
-				extKeysAndPubs.Add((ek, key));
-
-				if (ek.PrivateKey.PubKey.GetScriptPubKey(ScriptPubKeyType.Segwit) != key.P2wpkhScript
-					&& ek.PrivateKey.PubKey.GetScriptPubKey(ScriptPubKeyType.TaprootBIP86) != key.P2Taproot)
-				{
-					throw new InvalidOperationException("Wtf");
-				}
+				extKeysAndPubs.Add(ek);
 			}
 		}
 		return extKeysAndPubs;