--- a/WalletWasabi/Extensions/LazyExtensions.cs
+++ b/WalletWasabi/Extensions/LazyExtensions.cs
@@ -1,0 +1,7 @@
+namespace WalletWasabi.Extensions;
+
+public static class LazyExtensions
+{
+	public static Lazy<R> Select<T, R>(this Lazy<T> me, Func<T, R> map) =>
+		new (() => map(me.Value));
+}