This PR is for simplifying the `HdPubKeyCache` that contains unnecessary complexities that make it slow (the idea behind whatever is named blahblahCache is to make things faster.

The PR contains many things, changes and improvements all across the key management code because I think those are changes that we need to do anyway but by sure not all of them in the same PR.

One clear improvement in term of performance is in the `HdPubKeyCache::GetView` (a method that is called millions of times). 

## master (results in microseconds)
```
| Method  | WalletSize | Mean         | Error     | StdDev    |
|-------- |----------- |-------------:|----------:|----------:|
| GetView | 100        |     9.994 us | 0.0770 us | 0.0643 us |
| GetView | 1000       |    99.277 us | 1.2050 us | 1.0682 us |
| GetView | 10000      | 1,248.137 us | 5.1878 us | 4.5989 us |
```

## PR (result in nanoseconds 1us = 1000 ns)
```
| Method  | WalletSize | Mean      | Error    | StdDev   | Median    |
|-------- |----------- |----------:|---------:|---------:|----------:|
| GetView | 100        |  92.55 ns | 1.907 ns | 4.495 ns |  91.07 ns |
| GetView | 1000       |  92.21 ns | 1.218 ns | 1.139 ns |  91.91 ns |
| GetView | 10000      | 108.51 ns | 1.284 ns | 1.201 ns | 107.79 ns |
```

About the new brain fart involving `ArraySegment`, that's there to avoid changing behavior. I mean, I think it is absolutely unnecessary but it is to avoid thinking and testing.