@turbolay 2024-01-13
To help me reviewing, can you explain why a version using only a Dictionary in the HdPubKeyCache is not good, or can change behavior? https://github.com/turbolay/WalletWasabi/commit/b75679d03d41f4d4ffa1c2cf6ee14697186993d8