@lontivero 2024-01-14
Using only a dictionary is possible but the assumption in the code is that the `HdPubKeyCache` always gives you an `IEnumerable` over a collection that never changes, in master this is achieved by returning a clone converted to an `ImmutableList` while the version in this PR achieves the same by using an `ArraySegment` over an array that only adds up so, the question is: how do you achieve the same using a dictionary?