@turbolay 2024-01-14
Thank you this is exactly the answer I needed. I will review soon