@molnard 2024-01-17
I loaded an old wallet and I got the following exc. 

```
System.NullReferenceException
  HResult=0x80004003
  Message=Object reference not set to an instance of an object.
  Source=WalletWasabi
  StackTrace:
   at WalletWasabi.Blockchain.Keys.HdPubKeyCache.<>c.<get_HdPubKeys>b__6_0(HdPubKeyInfo x) in WalletWasabi\Blockchain\Keys\HdPubKeyCache.cs:line 19
```

![image](https://github.com/zkSNACKs/WalletWasabi/assets/9844978/49831fff-4003-4e50-b58a-07b559b4c96c)

![image](https://github.com/zkSNACKs/WalletWasabi/assets/9844978/edde68fc-2872-4e2b-a74c-6540064d271b)